fix: norminette, cylinder projection

+++ src/ray_tracing/intersection.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   intersection.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tklimova <tklimova@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tklimova <tklimova@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/20 22:48:37 by tklimova          #+#    #+#             */
-/*   Updated: 2024/05/10 15:20:35 by tklimova         ###   ########.fr       */
+/*   Updated: 2024/05/10 18:33:44 by tklimova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,7 +73,7 @@ t_closest_obj	get_closest_obj(t_mini_rt_data *data, t_ray ray)
 	t_closest_obj	closest_obj;
 	t_g_objects		*curr_obj;
 
-	closest_obj.dist = INT_MAX;
+	closest_obj.dist = (float) INT_MAX;
 	closest_obj.obj = NULL;
 	closest_obj.in_light = 0;
 	curr_obj = NULL;
